Add module to create a secondary Postgres login
This module will create a secondary Postgres login using supplied
grants. This is useful for creating read-only logins for reporting
tools.

The credentials are automatically rotated using Secrets Manager and
Lambda.

The function uses the alternating user mechanism:

- Two users are created: a primary and an alternate
- Each time the password rotates, it swaps the accounts and updates the
  password

If either user does not exist, it will be created and the specified
grants will be applied. Because it uses alternating users, the password
will remain active for twice the rotation period and applications won't
be disconnected when a rotation occurs.