Pass REPLICA_HOST address in as env var
The initial_values for the postgres login secrets doesn't seem to be
tracked as changes, so when the replica database host name is changes,
terraform doesn't recognize anything has changed. Olamide suggested
changing this module to pass replica host as an env var to the python
lambda rotation script to correct this.