Writes the RDS database URL under a different secret key for replicas
When creating the postgres login and the rotations for the replica, it
will create a DATABASE_URL secret under the new secret object. The
problem with that is that it will end up conflicting with the existing
DATABASE_URL. As a workaround, we've been configuring an object alias
for the secrets to avoid the conflict, which is not obvious and can
cause problems if someone forgets to do it.

This adds a boolean variable that tells whether the database is a
replica or not. When it is a replica, the key name should be
"REPLICA_DATABASE_URL", and when it isn't, it will keep the default
"DATABASE_URL".